Move mch264dec section in configure. More specific run-configure.sh.
Needed to move mch264dec to end since adding that framework without a -F
causes other test compiles to fail.

Changed run-configure.sh to explicitly enable/disable what we want
and had to fix to explicitly list libs for some components that depend
on autodetection to configure themselves (and we aren't autodetecting
we are explicitly enabling).